feat(LadderFactory): 높이와 폭을 가지는 LadderFactory 클래스 추가
